+++ source/desktop/screen_capturer_dxgi.cc
@@ -79,7 +79,7 @@ int indexFromScreenId(ScreenCapturer::ScreenId id, const std::vector<std::wstrin
 
     DCHECK_EQ(device_names.size(), screens.size());
 
-    for (int i = 0; i < screens.size(); ++i)
+    for (int i = 0; i < static_cast<int>(screens.size()); ++i)
     {
         if (screens[i].id == id)
             return i;
